+++ modules/landing/charts/OverallServiceChartWrapper.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { ChartPlaceHolder } from '../../../common/components/graphics/ChartPlaceHolder';
-import { useDailyTripsDataLanding } from '../../../common/api/hooks/dailytrips';
+import { useTripMetricsForLanding } from '../../../common/api/hooks/tripmetrics';
 import { OverallServiceChart } from './OverallServiceChart';
 
 export const OverallServiceChartWrapper: React.FC = () => {
-  const serviceData = useDailyTripsDataLanding();
+  const serviceData = useTripMetricsForLanding();
   const serviceDataReady = serviceData.some((query) => !query.isError && query.data);
   if (!serviceDataReady) return <ChartPlaceHolder query={serviceData[0]} />;
   const serviceDataFiltered = serviceData
