+++ modules/service/ServiceWidget.tsx
@@ -5,7 +5,7 @@ import { ChartPlaceHolder } from '../../common/components/graphics/ChartPlaceHol
 import { WidgetDiv } from '../../common/components/widgets/WidgetDiv';
 import { useTripCounts } from '../../common/api/hooks/service';
 import { OVERVIEW_OPTIONS, TODAY_STRING } from '../../common/constants/dates';
-import { useActualTripsDataByLine } from '../../common/api/hooks/dailytrips';
+import { useActualTripsDataByLine } from '../../common/api/hooks/tripmetrics';
 import { getSpeedGraphConfig } from '../speed/constants/speeds';
 import { HomescreenWidgetTitle } from '../dashboard/HomescreenWidgetTitle';
 import { ServiceGraphWrapper } from './ServiceGraphWrapper';
