+++ server/.chalice/policy.json
@@ -20,10 +20,9 @@
       "Action": ["dynamodb:Query"],
       "Effect": "Allow",
       "Resource": [
-        "arn:aws:dynamodb:*:*:table/DailySpeed",
-        "arn:aws:dynamodb:*:*:table/DailyTrips",
-        "arn:aws:dynamodb:*:*:table/WeeklySpeed",
-        "arn:aws:dynamodb:*:*:table/MonthlySpeed",
+        "arn:aws:dynamodb:*:*:table/DeliveredTripMetrics",
+        "arn:aws:dynamodb:*:*:table/DeliveredTripMetricsWeekly",
+        "arn:aws:dynamodb:*:*:table/DeliveredTripMetricsMonthly",
         "arn:aws:dynamodb:*:*:table/TripCounts",
         "arn:aws:dynamodb:*:*:table/Ridership"
       ]
