+++ server/app.py
@@ -170,7 +170,7 @@ def get_alerts():
     return json.dumps(response, indent=4, sort_keys=True, default=str)
 
 
-@app.route("/api/tripsbyline", cors=cors_config)
+@app.route("/api/tripmetrics", cors=cors_config)
 def get_trips_by_line():
     response = speed.trip_metrics_by_line(app.current_request.query_params)
     return json.dumps(response, indent=4, sort_keys=True)
